Move widget control panel app bar to separate component
